requirements.txt
dnspython
termcolor
retrying